+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # JCM v2.2.1 for MTR 4.0.5 has been released!
 
+This is a minor release. Server owners are advised to update to this version.
+
 ## Additions
 - MTR Patch: Skip rail culling for very large rail, to avoid visibility issue.
 - MTR Patch: Lower culling boundary calculation precision to help with lag spikes when creating large rail.
+
+## Changes
+- Add some extra permission guarding against block configuration packet.
 
 **Download:**  
 You can download this release on [Modrinth](https://modrinth.com/mod/jcm), [CurseForge](https://curseforge.com/minecraft/mc-mods/jcm) or [GitHub](https://github.com/DistrictOfJoban/Joban-Client-Mod/releases)